refactored PID Tracker, added logic in y-axis, split into modules
