+++ src/model_processors/BaseProcessor.py
@@ -1,7 +1,19 @@
 """
-Object factory for object-detection tasks - returns a fully initialized object 
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
 """
+
 import os 
+from abc import abstractmethod
 
 from atlas_utils.acl_resource import AclResource
 from atlas_utils.acl_model import Model
@@ -12,15 +24,28 @@ class BaseProcessor:
         self._acl_resource = AclResource()
         self._acl_resource.init()
         self.params = params
+        self.validate()
         self._model_width = params['model_width']
         self._model_height = params['model_height']
-        assert 'model_path' in params and params['model_path'] is not None, 'Review your param: model_path'
-        assert os.path.exists(params['model_path']), "Model directory doesn't exist {}".format(params['model_path'])
-        # Init Model
+        self.validate()
         self.model = Model(params['model_path'])
 
+    def validate(self):
+        if not os.path.exists(self.params['model_path']):
+            raise FileNotFoundError('Model Path not found, please check again.')
+        if 'model_width' not in self.params or 'model_height' not in self.params:
+            raise Exception('Please specify input width and height for model in params.py')
+
+    @abstractmethod
     def preprocess(self):
-        raise NotImplementedError
-    
+        pass
+        
+    @abstractmethod    
     def postprocess(self):
-        raise NotImplementedError
+        pass
+
+    @abstractmethod
+    def predict(self):
+        pass
+
+
